Update plan.md
We are running out of time on the implementation, and SkipList implementations are apparently rare to find out in the wild. For this reason, it is probably best to focus on queues and lists, and make sure that EBR and HP works. We should also take some time to improve the schemes, since comparisons of bad implementations isn't really interesting.